+++ services/tyre_model_service.py
@@ -2,9 +2,13 @@ import logging
 from sqlalchemy import or_
 from database.extensions import db
 from database.unit_of_work import UnitOfWork
+from policies.file_naming import uuid_filename
+from werkzeug.datastructures import FileStorage
 from database.models.tyre_model import TyreModel
-from domain.exceptions import DatabaseError, ModelNotFoundError
+from services.file_service import FileService, FileSaveRequest
+from database.models.data_types.files import FileType, FileModel
 from database.repositories.tyre_model_repository import TyreModelRepository
+from domain.exceptions import DatabaseError, ModelNotFoundError, InvalidFileTypeError, InvalidFileError, FileSaveError
 
 logger = logging.getLogger(__name__)
 
@@ -12,6 +16,7 @@ logger = logging.getLogger(__name__)
 class TyreModelService:
     def __init__(self):
         self.repo = TyreModelRepository()
+        self.file_service = FileService()
 
     def get_all(self, page=1, page_size=20, search="") -> (list[TyreModel], int):
         filters = None
@@ -26,14 +31,13 @@ class TyreModelService:
         return self.repo.get_all(page=page, page_size=page_size, filters=filters)
 
     def get_by_id(self, id_: int) -> TyreModel:
-        tyre_model =  self.repo.get_by_id(id_)
-
-        if not tyre_model:
-            logger.error(f"Error getting tyre model by id: {id_}")
-            raise ModelNotFoundError(f"Error getting tyre model by id: {id_}")
+        try:
+            tyre_model = self.repo.get_by_id(id_)
+        except ModelNotFoundError as e:
+            logger.error(f"TyreModel with id {id_} not found: {e}")
+            raise ModelNotFoundError(f"TyreModel with id {id_} not found: {e}") from e
 
         return tyre_model
-
 
     def create(self, dto) -> TyreModel:
         with UnitOfWork(db.session):
@@ -58,13 +62,50 @@ class TyreModelService:
 
         return tyre_model
 
+    def upload_image(self, tyre_model: TyreModel, file: FileStorage) -> TyreModel:
+        if not file:
+            logger.error("No file provided in tyre model service")
+            raise InvalidFileTypeError("No file provided")
+
+        uuid, filename = uuid_filename(file)
+        file.filename = filename
+
+        with UnitOfWork(db.session):
+            try:
+                file_record = self.file_service.handle_file(
+                    FileSaveRequest(
+                        file=file,
+                        upload_directory=f"/tyre_match/files/tyre_models/{tyre_model.id}",
+                        valid_extensions=["png", "jpg", "jpeg"],
+                        model=FileModel.tyre_model,
+                        model_id=tyre_model.id,
+                        file_type=FileType.original
+                    )
+                )
+            except InvalidFileError as e:
+                logger.error(f"Invalid file error from file service in tyre model service: {e}")
+                raise FileSaveError(f"Invalid file error from file service in tyre model service: {e}") from e
+            except InvalidFileTypeError as e:
+                logger.error(f"Invalid file type error from file service in tyre model service: {e}")
+                raise FileSaveError(f"Invalid file type error from file service in tyre model service: {e}") from e
+            except (PermissionError, OSError) as e:
+                logger.error(f"Permission or OS error from file service in tyre model service: {e}")
+                raise FileSaveError(f"Permission or OS error from file service in tyre model service: {e}") from e
+            except DatabaseError as e:
+                logger.error(f"Database error from file service in tyre model service: {e}")
+                raise FileSaveError(f"Database error from file service in tyre model service: {e}") from e
+
+            tyre_model.files.append(file_record)
+
+        return tyre_model
+
     def update(self, id_, dto) -> TyreModel:
         with UnitOfWork(db.session):
-            tyre_model = self.repo.get_by_id(id_)
-
-            if not tyre_model:
-                logger.error(f"Error getting tyre model by id: {id_}")
-                raise ModelNotFoundError(f"Error getting tyre model by id: {id_}")
+            try:
+                tyre_model = self.repo.get_by_id(id_)
+            except ModelNotFoundError as e:
+                logger.error(f"TyreModel with id {id_} not found: {e}")
+                raise ModelNotFoundError(f"TyreModel with id {id_} not found: {e}") from e
 
             try:
                 allowed_fields = [