+++ tests/services/test_tyre_model_service.py
@@ -1,14 +1,20 @@
 import pytest
 from unittest.mock import patch
+from services.file_service import FileService
+from werkzeug.datastructures import FileStorage
+from database.models.data_types.files import FileModel
 from services.tyre_model_service import TyreModelService
-from domain.exceptions import ModelNotFoundError, DatabaseError
 from tests.helpers.factories.tyre_model_factory import TyreModelFactory
 from database.repositories.tyre_model_repository import TyreModelRepository
+from domain.exceptions import ModelNotFoundError, DatabaseError, InvalidFileTypeError, InvalidFileError, FileSaveError
 
 
-def test_get_all():
-    service = TyreModelService()
+@pytest.fixture()
+def service():
+    return TyreModelService()
 
+
+def test_get_all(service):
     tyre_model_1 = TyreModelFactory().create()
     tyre_model_2 = TyreModelFactory().create()
     tyre_model_3 = TyreModelFactory().create()
@@ -21,9 +27,7 @@ def test_get_all():
     assert tyre_model_3 in results
 
 
-def test_get_all_pagination_page_1():
-    service = TyreModelService()
-
+def test_get_all_pagination_page_1(service):
     tyre_model_1 = TyreModelFactory().create()
     tyre_model_2 = TyreModelFactory().create()
     tyre_model_3 = TyreModelFactory().create()
@@ -37,9 +41,7 @@ def test_get_all_pagination_page_1():
     assert tyre_model_3 not in results
 
 
-def test_get_all_pagination_page_2():
-    service = TyreModelService()
-
+def test_get_all_pagination_page_2(service):
     tyre_model_1 = TyreModelFactory().create()
     tyre_model_2 = TyreModelFactory().create()
     tyre_model_3 = TyreModelFactory().create()
@@ -53,9 +55,7 @@ def test_get_all_pagination_page_2():
     assert tyre_model_3 not in results
 
 
-def test_get_all_search_by_manufacturer():
-    service = TyreModelService()
-
+def test_get_all_search_by_manufacturer(service):
     tyre_model_1 = TyreModelFactory().create(manufacturer='Test Manufacturer')
     tyre_model_2 = TyreModelFactory().create()
     tyre_model_3 = TyreModelFactory().create()
@@ -68,9 +68,7 @@ def test_get_all_search_by_manufacturer():
     assert tyre_model_3 not in results
 
 
-def test_get_all_search_by_model_name():
-    service = TyreModelService()
-
+def test_get_all_search_by_model_name(service):
     tyre_model_1 = TyreModelFactory().create(model_name='Test Model Name')
     tyre_model_2 = TyreModelFactory().create()
     tyre_model_3 = TyreModelFactory().create()
@@ -83,38 +81,28 @@ def test_get_all_search_by_model_name():
     assert tyre_model_3 not in results
 
 
-def test_get_by_id_invalid_id():
-    service = TyreModelService()
-
-    with pytest.raises(ModelNotFoundError, match="Error getting tyre model by id: 1"):
-        service.get_by_id(1)
+def test_get_by_id_invalid_id(service):
+    with pytest.raises(ModelNotFoundError, match="TyreModel with id 1000 not found"):
+        service.get_by_id(1000)
 
 
-def test_get_by_id():
-    service = TyreModelService()
-
+def test_get_by_id(service):
     tyre_model_1 = TyreModelFactory().create()
     tyre_model_2 = TyreModelFactory().create()
-    tyre_model_3 = TyreModelFactory().create()
 
     result = service.get_by_id(tyre_model_1.id)
 
     assert tyre_model_1 == result
     assert tyre_model_2 != result
-    assert tyre_model_3 != result
 
 
-def test_create_database_error_from_tyre_model_repository():
-    service = TyreModelService()
-
+def test_create_database_error_from_tyre_model_repository(service):
     with patch.object(TyreModelRepository, 'create', side_effect=DatabaseError("test error")):
         with pytest.raises(DatabaseError, match="Error creating tyre model record: test error"):
             service.create({})
 
 
-def test_create():
-    service = TyreModelService()
-
+def test_create(service):
     dto = {
         "manufacturer": "Test Manufacturer",
         "model_name": "Test Model Name",
@@ -127,16 +115,83 @@ def test_create():
     assert result.model_name == "Test Model Name"
 
 
-def test_update_invalid_id():
-    service = TyreModelService()
+def test_upload_image_no_file(service):
+    tyre_model = TyreModelFactory().create()
 
-    with pytest.raises(ModelNotFoundError, match="Error getting tyre model by id: 1"):
-        service.get_by_id(1)
+    with pytest.raises(InvalidFileTypeError, match="No file provided"):
+        service.upload_image(tyre_model, None)
 
 
-def test_update_database_error_from_tyre_model_repository():
-    service = TyreModelService()
+def test_upload_image_invalid_file_error_from_file_service(service):
+    tyre_model = TyreModelFactory().create()
 
+    file = FileStorage(filename="test-file.jpg")
+
+    with patch.object(FileService, "handle_file", side_effect=InvalidFileError("test error")):
+        with pytest.raises(FileSaveError, match="Invalid file error from file service in tyre model service: test error"):
+            service.upload_image(tyre_model, file)
+
+
+def test_upload_image_invalid_file_type_error_from_file_service(service):
+    tyre_model = TyreModelFactory().create()
+
+    file = FileStorage(filename="test-file.jpg")
+
+    with patch.object(FileService, "handle_file", side_effect=InvalidFileTypeError("test error")):
+        with pytest.raises(FileSaveError, match="Invalid file type error from file service in tyre model service: test error"):
+            service.upload_image(tyre_model, file)
+
+
+def test_upload_image_permission_error_from_file_service(service):
+    tyre_model = TyreModelFactory().create()
+
+    file = FileStorage(filename="test-file.jpg")
+
+    with patch.object(FileService, "handle_file", side_effect=PermissionError("test error")):
+        with pytest.raises(FileSaveError, match="Permission or OS error from file service in tyre model service: test error"):
+            service.upload_image(tyre_model, file)
+
+
+def test_upload_image_os_error_from_file_service(service):
+    tyre_model = TyreModelFactory().create()
+
+    file = FileStorage(filename="test-file.jpg")
+
+    with patch.object(FileService, "handle_file", side_effect=OSError("test error")):
+        with pytest.raises(FileSaveError, match="Permission or OS error from file service in tyre model service: test error"):
+            service.upload_image(tyre_model, file)
+
+
+def test_upload_image_database_error_from_file_service(service):
+    tyre_model = TyreModelFactory().create()
+
+    file = FileStorage(filename="test-file.jpg")
+
+    with patch.object(FileService, "handle_file", side_effect=DatabaseError("test error")):
+        with pytest.raises(FileSaveError, match="Database error from file service in tyre model service: test error"):
+            service.upload_image(tyre_model, file)
+
+
+def test_upload_image(service):
+    tyre_model = TyreModelFactory().create()
+
+    file = FileStorage(filename="test-file.jpg")
+
+    res = service.upload_image(tyre_model, file)
+
+    assert res is not None
+    assert res.id == tyre_model.id
+    assert 1 == len(res.files)
+    assert res.files[0].model == FileModel.tyre_model
+    assert res.files[0].model_id == tyre_model.id
+
+
+def test_update_invalid_id(service):
+    with pytest.raises(ModelNotFoundError, match="TyreModel with id 1000 not found"):
+        service.get_by_id(1000)
+
+
+def test_update_database_error_from_tyre_model_repository(service):
     tyre_model = TyreModelFactory().create()
 
     with patch.object(TyreModelRepository, 'update', side_effect=DatabaseError("test error")):
@@ -144,9 +199,7 @@ def test_update_database_error_from_tyre_model_repository():
             service.update(tyre_model.id, {})
 
 
-def test_update():
-    service = TyreModelService()
-
+def test_update(service):
     tyre_model = TyreModelFactory().create()
 
     updated_model = service.update(
@@ -162,7 +215,7 @@ def test_update():
     assert "Test Model Name" == updated_model.model_name
 
 
-def test_delete_invalid_id():
+def test_delete_invalid_id(service):
     service = TyreModelService()
 
     with pytest.raises(ModelNotFoundError, match="Tyre model with id 1 not found"):
@@ -170,22 +223,18 @@ def test_delete_invalid_id():
         assert result == False
 
 
-def test_delete_database_error_from_tyre_model_repository():
-    service = TyreModelService()
-
+def test_delete_database_error_from_tyre_model_repository(service):
     with patch.object(TyreModelRepository, 'delete', side_effect=DatabaseError("test error")):
         with pytest.raises(DatabaseError, match="Error deleting tyre model record: test error"):
             result = service.delete(1)
             assert result == False
 
 
-def test_delete():
-    service = TyreModelService()
-
+def test_delete(service):
     tyre_model = TyreModelFactory().create()
 
     result = service.delete(tyre_model.id)
 
     assert result == True
-    with pytest.raises(ModelNotFoundError, match=f"Error getting tyre model by id: {tyre_model.id}"):
+    with pytest.raises(ModelNotFoundError, match=f"TyreModel with id {tyre_model.id} not found"):
         service.get_by_id(tyre_model.id)